% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detalles}
\alias{detalles}
\title{Descripci칩n de los par치metros}
\format{
\subsection{\code{detalles}}{

Una lista con los detalles y opciones de cada par치metro.
}
}
\source{
\url{https://desarrollador.sb.gob.do/referencias#uk1jx}
}
\usage{
detalles
}
\description{
Se hizo un webscraping de la documentaci칩n del API de la superintendencia de
bancos. \url{https://desarrollador.sb.gob.do/referencias#uk1jx}
}
\keyword{datasets}
